Organize PlantGrowthChamber into a folder
