+++ src/problems/PlantGrowthChamberBinary.py
@@ -1,6 +1,6 @@
 from PyExpUtils.collection.Collector import Collector
 
-from environments.PlantGrowthChamberIntensity import PlantGrowthChamberIntensity as Env
+from environments.PlantGrowthChamber.PlantGrowthChamberIntensity import PlantGrowthChamberIntensity as Env
 from experiment.ExperimentModel import ExperimentModel
 from problems.BaseProblem import BaseProblem
